+++ benchmarks/verus/dafnybench/files/cmsc433_tmp_tmpe3ob3a0o_dafny_project1_p1-assignment-2_IsPrime.rs
@@ -8,9 +8,9 @@ verus! {
 // </vc-helpers>
 
 // <vc-spec>
-fn is_prime(m: i32) -> (result: bool)
-    requires m > 0
-    ensures result <==> (m > 1 && forall|j: int| 2 <= j < m ==> #[trigger] (m % j as i32) != 0)
+fn is_prime(m: i8) -> (result: bool)
+    requires m as int > 0
+    ensures result <==> (m as int > 1 && forall|j: int| 2 <= j < m as int ==> #[trigger] (m as int % j) != 0)
 // </vc-spec>
 // <vc-code>
 {
