+++ benchmarks/verus/dafnybench/files/dafny-synthesis_task_id_127_Multiply.rs
@@ -8,8 +8,8 @@ verus! {
 // </vc-helpers>
 
 // <vc-spec>
-fn multiply(a: int, b: int) -> (result: int)
-    ensures result == a * b
+fn multiply(a: i8, b: i8) -> (result: i8)
+    ensures result as int == a as int * b as int
 // </vc-spec>
 // <vc-code>
 {
