+++ benchmarks/verus/dafnybench/files/dafny-synthesis_task_id_581_SquarePyramidSurfaceArea.rs
@@ -8,12 +8,12 @@ verus! {
 // </vc-helpers>
 
 // <vc-spec>
-fn square_pyramid_surface_area(base_edge: i32, height: i32) -> (area: i32)
+fn square_pyramid_surface_area(base_edge: i8, height: i8) -> (area: i8)
     requires 
         base_edge > 0,
         height > 0,
     ensures 
-        area == base_edge * base_edge + 2 * base_edge * height,
+        area as int == base_edge as int * base_edge as int + 2 * base_edge as int * height as int,
 // </vc-spec>
 // <vc-code>
 {
