+++ benchmarks/verus/dafnybench/files/formal_verication_dafny_tmp_tmpwgl2qz28_Challenges_ex2_Allow42.rs
@@ -8,10 +8,10 @@ verus! {
 // </vc-helpers>
 
 // <vc-spec>
-fn allow_42(x: i32, y: i32) -> (ret: (i32, bool))
+fn allow_42(x: i8, y: i8) -> (ret: (i8, bool))
     ensures 
-        (y != 42 ==> ret.0 == (x as int / (42int - y as int)) as i32 && ret.1 == false) &&
-        (y == 42 ==> ret.0 == 0 && ret.1 == true)
+        (y as int != 42 ==> ret.0 as int == (x as int / (42int - y as int)) && ret.1 == false) &&
+        (y as int == 42 ==> ret.0 as int == 0 && ret.1 == true)
 // </vc-spec>
 // <vc-code>
 {
